xpp: refactor simd files
