+++ libxpp/simd/XppSimd_x86.c
